update: move interface to sub module
